refactor: Extract the categories into its own component.
