+++ packages/map-template/src/atoms/searchResultsState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const searchResultsState = atom({
+    key: 'searchResults',
+    default: []
+});
+
+export default searchResultsState;
