+++ packages/map-template/src/atoms/selectedCategoryState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const selectedCategoryState = atom({
+    key: 'selectedCategory',
+    default: undefined
+});
+
+export default selectedCategoryState;
