+++ packages/map-template/src/components/Categories/Categories.scss
@@ -1,0 +1,29 @@
+.categories {
+    overflow: auto;
+    display: grid;
+    grid-auto-flow: column;
+    align-items: end;
+    gap: var(--spacing-x-small);
+
+    &__list {
+        display: grid;
+        grid-auto-flow: column;
+        gap: var(--spacing-x-small);
+        overflow: auto hidden;
+        grid-auto-columns: min-content;
+        margin-top: var(--spacing-x-small);
+    }
+
+    &__scroll-button {
+        width: var(--spacing-xxx-large);
+        height: var(--spacing-xxx-large);
+        background-color: var(--brand-colors-white-100);
+        border: 1px solid var(--color-gray-30);
+        border-radius: var(--spacing-x-small);
+
+        &:disabled {
+            cursor: not-allowed;
+            opacity: 0.3;
+        }
+    }
+}
